Python script to selectively disable keys according to focus
Currently this just watches for Firefox's pop-out media player and
disables the Escape key when it's active.

This is useful because I use a vi like text editor (Helix) and I have a
habit of hitting Esc everywhere.  This causes the video player to close
which isn't what I want to do (I can always use my real keybinding for
closing a window if I really want to close it!).