Disable escape in firefox player (#13)
* Python script to selectively disable keys according to focus

Currently this just watches for Firefox's pop-out media player and
disables the Escape key when it's active.

This is useful because I use a vi like text editor (Helix) and I have a
habit of hitting Esc everywhere.  This causes the video player to close
which isn't what I want to do (I can always use my real keybinding for
closing a window if I really want to close it!).

* add module level description
* use correct logger

* Make it easier to add other binds

Not sure who'd actually want to use it, but it's enough to add calls
like:

   mon.bind("Space", app_id="foot")

to prevent spacebar from being passed through to the foot terminal
emulator.