Improve the robustness of inactive-windows-transparency
Only storing the window and workspace that were focused immediately
prior is not enough to handle all cases. For example, denoting by A1 the
1st window on workspace A, if we switch focus from A1 to B1 both A1 and
B1 will be opaque. If we then switch to A2, our prev_focused is B1 which
will remain opaque as its on a different workspace. The end result is
that A1, A2, and B1 are opaque, leaving two opaque windows on the same
workspace.

This commit switches to storing a set of all windows that are opaque. We
don't store their workspaces to attempt to be more robust if they are
moved. On a focus event, we walk this set, making windows that are on
the same workspace as the newly focused window translucent and removing
them from the set.