Fix grimshot trying to send notification even when disabled
Grimshot was trying to send notification by default, even when not
enabled.

The bug was that the check NOTIFY=no was done in the when() function.
The "enter" was thus exiting the when() function instead of the
notifyOk().

Fix was to replace the when() call by a simple if/then/fi